+++ keyboards/system76/system76_ec.c
@@ -144,8 +144,8 @@ rgb_config_t layer_rgb[DYNAMIC_KEYMAP_LAYER_COUNT] = {
 #ifdef SYSTEM76_EC
 // Read or write EEPROM data with checks for being inside System76 EC region.
 static bool system76_ec_eeprom_op(void *buf, uint16_t size, uint16_t offset, bool write) {
-    uint16_t addr = SYSTEM76_EC_EEPROM_ADDR + offset;
-    uint16_t end  = addr + size;
+    uintptr_t addr = SYSTEM76_EC_EEPROM_ADDR + offset;
+    uintptr_t end  = addr + size;
     // Check for overflow and zero size
     if ((end > addr) && (addr >= SYSTEM76_EC_EEPROM_ADDR) && (end <= (SYSTEM76_EC_EEPROM_ADDR + SYSTEM76_EC_EEPROM_SIZE))) {
         if (write) {
